Saves institution-specific stuff in tmp location
This commit moves IU- and WGBH-specific files in a tmp location. After the files
are removed from master, and this branch is rebased onto master, then these
files will be moved back to their normal location, and they'll only exist in the
IU- and WGBH-specific branches.

This is an intermediary step to separating the code bases altogether. Eventually
we want all HydraDAM functionality in a gem that can be used to augment a
CurationConcerns app. Then each institution will have their own CurationConcerns
app that uses HydraDAM. Any institution-specific customization will happen
within the repsective repositories. However, we're holding off on doing that
at this time.