Sets the title of an ingested work
Previously, it wasn't getting set, so the display value was just the Fedora ID,
which isn't useful. I set it to the Barcode from the MDPI xml. This seems to be
what all other file names are based on. But we can change it if that's not what
we want.